requirements.txt
requests~=2.32.3
flask~=3.0.3
androguard~=4.1.2
pymongo~=4.8.0
pandas~=2.2.2
beautifulsoup4~=4.12.3
zhipuai~=2.1.2
flask_cors~=4.0.0
pyzbar~=0.1.9
pillow~=10.4.0